remove libft

+++ srcs/malloc.c
@@ -1,5 +1,4 @@
 #include "malloc.h"
-#include "libft.h"
 
 void		print_zone(int zone)
 {
@@ -24,10 +23,10 @@ size_t		print_chunk(t_header **block, int zone)
 	{
 		zone_size += tmp->size;
 		ft_putstr("0x");
-		ft_putnbr_base(tmp->mem, BASE16);
+		ft_putnbr_base((int)tmp->mem, BASE16);
 		ft_putstr(" - ");
 		ft_putstr("0x");
-		ft_putnbr_base(tmp->mem + tmp->size, BASE16);
+		ft_putnbr_base((int)tmp->mem + tmp->size, BASE16);
 		ft_putstr(" : ");
 		ft_putnbr(tmp->size);
 		ft_putendl(" octets");
